Create a bar chart with scores per criteria cluster for every alternative in a decision

Create a bar chart with scores per criteria cluster for every alternative in a decision

scoreBarchart_criteriaCluster_singleAlternative(
  multiEstimateDf,
  estimateCol,
  criteria,
  decision_id,
  alternative_value,
  parentCriterionOrder = unique(weightedEstimates$parentCriterion_id),
  parentCriterionLabels = parentCriterion_ids,
  decisionLabels = NULL,
  alternativeLabels = NULL,
  wrapLabels = 100,
  scoreRange = NULL,
  fillColor = "black",
  strokeColor = "black",
  strokeSize = 0.1,
  flipCoordinates = TRUE,
  title = "MDMCDA criteria cluster bar chart by alternative",
  xLab = "Criteria clusters",
  yLab = estimateCol,
  theme = ggplot2::theme_minimal(base_size = mdmcda::opts$get("ggBaseSize")),
  guides = ggplot2::guide_legend(nrow = 1),
  legend.position = "top"
)

Arguments

multiEstimateDf
A `multiEstimateDf` data frame.

estimateCol
The column name with the estimates to use.

criteria
A `criteria` object.

decision_id
The identifier of the decision to show.

alternative_value
The alternative value to show.

parentCriterionOrder
The parent criteria to use, and their order.

decisionLabels
The labels for the decisions (a named vector).

alternativeLabels
The labels for the alternatives (a list of named vectors).

wrapLabels
Where to wrap labels (`NULL` to not wrap).

fillColor, strokeColor, strokeSize
The color and pen width of the fill and stroke.

flipCoordinates
Whether to flip the coordinates.

title, xLab, yLab
The title and x and y axis labels.

theme
The `ggplot2` theme to use.

guides
A guides argument to tweak the legend.

legend.position, legend.box.margin
The position and spacing for the legend.

parentCriterion_labels
The labels for the parent criteria (a named vector).

Value

A `ggplot2` plot.