Create a bar chart with scores per criteria cluster per scenario

Create a bar chart with scores per criteria cluster per scenario

scoreBarchart_scenariosByCluster(
  weighedEstimates,
  estimateCol,
  parentCriterion_ids = unique(weighedEstimates$parentCriterion_id),
  parentCriterion_labels = NULL,
  scenario_ids = unique(weighedEstimates$scenario_id),
  scenario_labels = NULL,
  strokeColor = "black",
  strokeSize = 0.1,
  title = "MDMCDA scenarios by criteria cluster bar chart",
  xLab = "Scenario",
  yLab = estimateCol,
  theme = ggplot2::theme_minimal(base_size = mdmcda::opts$get("ggBaseSize")),
  guides = ggplot2::guide_legend(nrow = 2),
  legend.position = "top"
)

Arguments

weighedEstimates
A `weighedEstimates` object.

estimateCol
The column name with the estimates to use.

parentCriterion_ids
The parent criteria to include.

parentCriterion_labels
The labels for the parent criteria.

scenario_ids
The scenarios to include.

scenario_labels
The labels for the scenarios.

strokeColor, strokeSize
The color and pen width of the stroke.

title, xLab, yLab
The title and x and y axis labels.

theme
The `ggplot2` theme to use.

guides
A guides argument to tweak the legend.

legend.position, legend.box.margin
The position and spacing for the legend.

Value

A `ggplot2` plot.